requirements.txt
asn1crypto==0.24.0
attrs==19.1.0
backcall==0.1.0
bleach==3.1.0
Bottleneck==1.2.1
certifi==2019.3.9
cffi==1.12.2
chardet==3.0.4
Click==7.0
cloudpickle==0.8.0
cryptography==2.6.1
cycler==0.10.0
cymem==2.0.2
cytoolz==0.9.0.1
dask==1.1.4
dataclasses==0.6
decorator==4.3.2
defusedxml==0.5.0
dill==0.2.9
entrypoints==0.3
fastai==1.0.39
fastprogress==0.1.20
Flask==1.0.2
future==0.17.1
idna==2.8
ipykernel==5.1.0
ipython==7.3.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
itsdangerous==1.1.0
jedi==0.13.3
Jinja2==2.10
jsonschema==3.0.1
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.4.0
jupyterlab==0.35.4
jupyterlab-server==0.2.0
kiwisolver==1.0.1
lmdb==0.94
MarkupSafe==1.1.1
matplotlib==3.0.2
mistune==0.8.4
msgpack==0.5.6
msgpack-numpy==0.4.3.2
murmurhash==1.0.2
nbconvert==5.4.1
nbformat==4.4.0
networkx==2.2
notebook==5.7.6
numexpr==2.6.9
numpy==1.15.4
nvidia-ml-py3==7.352.0
opencv-python==4.0.0.21
packaging==19.0
pandas==0.23.4
pandocfilters==1.4.2
parso==0.3.4
pexpect==4.6.0
pickleshare==0.7.5
Pillow==5.3.0
pipdeptree==0.13.2
plac==0.9.6
preshed==2.0.1
prometheus-client==0.6.0
prompt-toolkit==2.0.9
protobuf==3.6.1
ptyprocess==0.6.0
pycparser==2.19
Pygments==2.3.1
pyparsing==2.3.1
pyrsistent==0.14.11
python-dateutil==2.8.0
python-telegram-bot==11.1.0
pytz==2018.9
PyWavelets==1.0.2
PyYAML==3.13
pyzmq==18.0.1
qtconsole==4.4.3
regex==2018.1.10
requests==2.21.0
scikit-image==0.14.2
scikit-video==1.1.11
scipy==1.2.0
Send2Trash==1.5.0
six==1.12.0
spacy==2.0.18
tensorboardX==1.6
terminado==0.8.1
testpath==0.4.2
thinc==6.12.1
toolz==0.9.0
torch==1.0.1.post2
torchvision==0.2.1
tornado==5.1.1
tqdm==4.28.1
traitlets==4.3.2
typing==3.6.6
ujson==1.35
urllib3==1.24.1
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
widgetsnbextension==3.4.2
wrapt==1.10.11
